Summa addition (#243)
* added SUMMA config generation to modular config fn

* tests

* Modular refactor (#244)

* remove duplicate code

* handle calibrated parameters

* remove model type name field, not used in ngen

* reorganize functions

* refactor tests

* remove redundant tests

* remove unused model cli args

* remove awkward prompt placement

* document calibrated parameter fetching

* aggregate new model info into one place

* remove hardcoded variable maps from templates

* update stale comments

* clarifying comments

* pulled out awkward rule

* remove confusing variable

* add this branch to ci
