+++ .github/workflows/ci.yml
@@ -2,7 +2,7 @@ name: CI
 
 on:
   pull_request:
-    branches: [main, modular-configs, casam, other-realizations]
+    branches: [main, modular-configs, summa-addition]
 
 concurrency:
   group: ci-${{ github.ref }}
