+++ modules/data_processing/create_realization.py
@@ -1,857 +1,569 @@
-import json
-import logging
-import multiprocessing
-import os
-import shutil
-import sqlite3
-from datetime import datetime
-from itertools import chain
-from pathlib import Path
-from typing import Dict, Optional
+"""Module to generate modular realizations."""
 
-import duckdb
-import numpy as np
-import pandas
-import psutil
+import copy
+import json
+from datetime import datetime
+import logging
+from dataclasses import dataclass, field
+from typing import Optional
+from pathlib import Path
 import requests
-import s3fs
-import xarray as xr
-from data_processing.dask_utils import temp_cluster
+
 from data_processing.file_paths import FilePaths
-from data_processing.gpkg_utils import get_cat_to_nhd_feature_id, get_table_crs_short
-from pyproj import Transformer
-from tqdm.rich import tqdm
 
 logger = logging.getLogger(__name__)
 
+@dataclass(frozen=True)
+class ModelSpec:
+    """All per-model coupling knowledge for one model.
 
-@temp_cluster
-def get_approximate_gw_storage(paths: FilePaths, start_date: datetime) -> Dict[str, int]:
-    # get the gw levels from the NWM output on a given start date
-    # this kind of works in place of warmstates for now
-    year = start_date.strftime("%Y")
-    formatted_dt = start_date.strftime("%Y%m%d%H%M")
-    cat_to_feature = get_cat_to_nhd_feature_id(paths.geopackage_path)
+    main_output_variable:
+        The module's terminal BMI output. This is what would get passed to t-route.
+    realization_fragment:
+        Path to the modular-realization JSON stub for this model. ``None`` means
+        the model is recognized but not yet wired end-to-end — ``supported`` is
+        ``False`` and ``validate_models`` rejects it with a clear message rather
+        than letting it KeyError deep in realization assembly.
+    routable:
+        True if it is a rainfall-runoff model whose outputs can be passed to t-route.
+        False if not.
+    variables_names_map:
+        Default variables_names_map assuming NO coupling. Empty for models that
+        carry none (LSTM, SLoTH).
+    overrides:
+        (upstream_provider, {input_name: source_name}) entries applied when the
+        provider precedes this model in the run order.
+    """
 
-    fs = s3fs.S3FileSystem(anon=True)
-    nc_url = f"s3://noaa-nwm-retrospective-3-0-pds/CONUS/netcdf/GWOUT/{year}/{formatted_dt}.GWOUT_DOMAIN1"
-
-    with fs.open(nc_url) as file_obj:
-        ds = xr.open_dataset(file_obj)  # type: ignore
-
-        water_levels: Dict[str, int] = dict()
-        for cat, feature in tqdm(cat_to_feature.items()):
-            # this value is in CM, we need meters to match max_gw_depth
-            # xarray says it's in mm, with 0.1 scale factor. calling .values doesn't apply the scale
-            water_level = ds.sel(feature_id=feature).depth.values / 100
-            water_levels[cat] = water_level
-
-    return water_levels
-
-
-def make_cfe_config(divide_conf_df: pandas.DataFrame, files: FilePaths, water_levels: dict) -> None:
-    """Parses parameters from NOAHOWP_CFE DataFrame and returns a dictionary of catchment configurations."""
-    with open(FilePaths.template_cfe_config, "r") as f:
-        cfe_template = f.read()
-    cat_config_dir = files.config_dir / "cat_config" / "CFE"
-    cat_config_dir.mkdir(parents=True, exist_ok=True)
-
-    for _, row in divide_conf_df.iterrows():
-        nwm_water_level = water_levels.get(row["divide_id"], None)
-        # if we have the nwm output water level for that catchment, use it
-        # otherwise, use 5%
-        if nwm_water_level is not None:
-            gw_storage_ratio = water_levels[row["divide_id"]] / row["mean.Zmax"]
-        else:
-            gw_storage_ratio = 0.05
-        cat_config = cfe_template.format(
-            bexp=row["mode.bexp_soil_layers_stag=2"],
-            dksat=row["geom_mean.dksat_soil_layers_stag=2"],
-            psisat=row["geom_mean.psisat_soil_layers_stag=2"],
-            slope=row["mean.slope_1km"],
-            smcmax=row["mean.smcmax_soil_layers_stag=2"],
-            smcwlt=row["mean.smcwlt_soil_layers_stag=2"],
-            max_gw_storage=row["mean.Zmax"] / 1000
-            if row["mean.Zmax"] is not None
-            else "0.011[m]",  # mean.Zmax is in mm!
-            gw_Coeff=row["mean.Coeff"] if row["mean.Coeff"] is not None else "0.0018[m h-1]",
-            gw_Expon=row["mode.Expon"],
-            gw_storage="{:.5}".format(gw_storage_ratio),
-            refkdt=row["mean.refkdt"],
-        )
-        cat_ini_file = cat_config_dir / f"{row['divide_id']}.ini"
-        with open(cat_ini_file, "w") as f:
-            f.write(cat_config)
+    main_output_variable: str
+    realization_fragment: Path | None
+    routable: Optional[bool] = False
+    variables_names_map: dict[str, str] = field(default_factory=dict)
+    overrides: list[tuple[str, dict[str, str]]] = field(default_factory=list)
 
 
-def make_noahowp_config(
-    base_dir: Path, divide_conf_df: pandas.DataFrame, start_time: datetime, end_time: datetime
-) -> None:
-    start_datetime = start_time.strftime("%Y%m%d%H%M")
-    end_datetime = end_time.strftime("%Y%m%d%H%M")
-    with open(FilePaths.template_noahowp_config, "r") as file:
-        template = file.read()
+# ---------------------------------------------------------------------------
+# THE REGISTRY
+# One entry per model. Order here is only for readability.
+# ---------------------------------------------------------------------------
 
-    cat_config_dir = base_dir / "cat_config" / "NOAH-OWP-M"
-    cat_config_dir.mkdir(parents=True, exist_ok=True)
+MODEL_REGISTRY: dict[str, ModelSpec] = {
+    "sloth": ModelSpec(
+        main_output_variable="z",
+        realization_fragment=FilePaths.sloth_modular_config,
+    ),
+    "nom": ModelSpec(
+        main_output_variable="EVAPOTRANS",
+        realization_fragment=FilePaths.nom_modular_config,
+        variables_names_map={
+            "PRCPNONC": "precip_rate",
+            "Q2": "SPFH_2maboveground",
+            "SFCTMP": "TMP_2maboveground",
+            "UU": "UGRD_10maboveground",
+            "VV": "VGRD_10maboveground",
+            "LWDN": "DLWRF_surface",
+            "SOLDN": "DSWRF_surface",
+            "SFCPRS": "PRES_surface",
+        },
+    ),
+    "cfe": ModelSpec(
+        main_output_variable="Q_OUT",
+        realization_fragment=FilePaths.cfe_modular_config,
+        routable=True,
+        variables_names_map={
+            "atmosphere_water__liquid_equivalent_precipitation_rate": "APCP_surface",
+            "water_potential_evaporation_flux": "sloth_pet",
+            "ice_fraction_schaake": "sloth_ice_fraction_schaake",
+            "ice_fraction_xinanjiang": "sloth_ice_fraction_xinanjiang",
+            "soil_moisture_profile": "sloth_soil_moisture_profile",
+        },
+        overrides=[
+            (
+                "nom",
+                {
+                    "atmosphere_water__liquid_equivalent_precipitation_rate": "QINSUR",
+                    "water_potential_evaporation_flux": "EVAPOTRANS",
+                },
+            ),
+            ("snow17", {"atmosphere_water__liquid_equivalent_precipitation_rate": "raim"}),
+            # ("pet", {"water_potential_evaporation_flux": "water_potential_evaporation_flux"}),
+            # (
+            #     "sft",
+            #     {
+            #         "ice_fraction_schaake": "ice_fraction_schaake",
+            #         "ice_fraction_xinanjiang": "ice_fraction_xinanjiang",
+            #     },
+            # ),
+            # ("smp", {"soil_moisture_profile": "soil_moisture_profile"}),
+        ],
+    ),
+    "casam": ModelSpec(
+        main_output_variable="total_discharge",
+        realization_fragment=FilePaths.casam_modular_config,
+        routable=True,
+        variables_names_map={
+            "precipitation_rate": "precip_rate",
+            "potential_evapotranspiration_rate": "sloth_pet",
+            "soil_temperature_profile": "sloth_soil_temperature_profile",
+        },
+        overrides=[
+            ("nom", {"potential_evapotranspiration_rate": "EVAPOTRANS"}),
+            ("snow17", {"precipitation_rate": "raim"}),
+            # ("pet", {"potential_evapotranspiration_rate": "water_potential_evaporation_flux"}),
+            # ("sft", {"soil_temperature_profile": "soil_temperature_profile"}),
+        ],
+    ),
+    "snow17": ModelSpec(
+        main_output_variable="raim",
+        realization_fragment=FilePaths.snow17_modular_config,
+        variables_names_map={
+            "precip": "atmosphere_water__liquid_equivalent_precipitation_rate",
+            "tair": "land_surface_air__temperature",
+        },
+    ),
+    "sac-sma": ModelSpec(
+        main_output_variable="tci",
+        realization_fragment=FilePaths.sac_modular_config,
+        routable=True,
+        variables_names_map={
+            "precip": "atmosphere_water__liquid_equivalent_precipitation_rate",
+            "tair": "land_surface_air__temperature",
+            "pet": "sloth_pet",
+        },
+        overrides=[
+            ("nom", {"pet": "EVAPOTRANS"}),
+            ("snow17", {"precip": "raim"}),
+            # ("pet", {"pet": "water_potential_evaporation_flux"}),
+        ],
+    ),
+    "lstm": ModelSpec(
+        main_output_variable="land_surface_water__runoff_depth",
+        realization_fragment=FilePaths.lstm_modular_config,
+        routable=True
+    ),
+    "lstm_rust": ModelSpec(
+        main_output_variable="land_surface_water__runoff_depth",
+        realization_fragment=FilePaths.lstm_rust_modular_config,
+        routable=True
+    ),
+    "dhbv2": ModelSpec(
+        main_output_variable="land_surface_water__runoff_volume_flux",
+        realization_fragment=FilePaths.dhbv2_modular_config,
+        routable=True,
+        variables_names_map={
+            "atmosphere_water__liquid_equivalent_precipitation_rate": "precip_rate",
+            "land_surface_air__temperature": "TMP_2maboveground",
+            "atmosphere_air_water~vapor__relative_saturation": "SPFH_2maboveground",
+            "land_surface_radiation~incoming~longwave__energy_flux": "DLWRF_surface",
+            "land_surface_radiation~incoming~shortwave__energy_flux": "DSWRF_surface",
+            "land_surface_air__pressure": "PRES_surface",
+            "land_surface_wind__x_component_of_velocity": "UGRD_10maboveground",
+            "land_surface_wind__y_component_of_velocity": "VGRD_10maboveground",
+            "land_surface_water__runoff_volume_flux": "streamflow",
+        },
+    ),
+    "dhbv2_daily": ModelSpec(
+        main_output_variable="land_surface_water__runoff_volume_flux",
+        realization_fragment=FilePaths.dhbv2_daily_modular_config,
+        routable=True,
+        variables_names_map={
+            "atmosphere_water__liquid_equivalent_precipitation_rate": "precip_rate",
+            "land_surface_air__temperature": "TMP_2maboveground",
+            "atmosphere_air_water~vapor__relative_saturation": "SPFH_2maboveground",
+            "land_surface_radiation~incoming~longwave__energy_flux": "DLWRF_surface",
+            "land_surface_radiation~incoming~shortwave__energy_flux": "DSWRF_surface",
+            "land_surface_air__pressure": "PRES_surface",
+            "land_surface_wind__x_component_of_velocity": "UGRD_10maboveground",
+            "land_surface_wind__y_component_of_velocity": "VGRD_10maboveground",
+            "land_surface_water__runoff_volume_flux": "streamflow",
+        },
+    ),
+    "summa": ModelSpec(
+        main_output_variable="land_surface_water__runoff_volume_flux",
+        realization_fragment=FilePaths.summa_modular_config,
+        routable=True,
+        variables_names_map={
+            "atmosphere_water__precipitation_mass_flux": "precip_rate",
+            "land_surface_air__temperature": "TMP_2maboveground",
+            "atmosphere_air_water~vapor__relative_saturation": "SPFH_2maboveground",
+            "land_surface_wind__x_component_of_velocity": "UGRD_10maboveground",
+            "land_surface_wind__y_component_of_velocity": "VGRD_10maboveground",
+            "land_surface_radiation~incoming~shortwave__energy_flux": "DSWRF_surface",
+            "land_surface_radiation~incoming~longwave__energy_flux": "DLWRF_surface",
+            "land_surface_air__pressure": "PRES_surface",
+        },
+    ),
+    # ------------------------------------------------------------------
+    # Recognized but not yet wired: no realization realization_fragment, so supported
+    # is False and validate_models rejects them up front. Their names-maps
+    # and overrides are kept as scaffolding for when they are wired.
+    # ------------------------------------------------------------------
+    # "pet": ModelSpec(
+    #     main_output_variable="water_potential_evaporation_flux",
+    #     realization_fragment=None,
+    #     variables_names_map={"water_potential_evaporation_flux": "potential_evapotranspiration"},
+    # ),
+    # "sft": ModelSpec(
+    #     main_output_variable="num_cells",
+    #     realization_fragment=None,
+    #     variables_names_map={
+    #         "ground_temperature": "sloth_ground_temperature",
+    #         "soil_moisture_profile": "sloth_soil_moisture_profile",
+    #     },
+    #     overrides=[
+    #         ("nom", {"ground_temperature": "TGS"}),
+    #         ("smp", {"soil_moisture_profile": "soil_moisture_profile"}),
+    #     ],
+    # ),
+    # "smp": ModelSpec(
+    #     main_output_variable="soil_storage",
+    #     realization_fragment=None,
+    #     variables_names_map={
+    #         "soil_storage": "sloth_soil_storage",
+    #         "soil_storage_change": "sloth_soil_storage_change",
+    #         "num_wetting_fronts": "sloth_num_wetting_fronts",
+    #         "soil_moisture_wetting_fronts": "sloth_soil_moisture_wetting_fronts",
+    #         "soil_depth_wetting_fronts": "sloth_soil_depth_wetting_fronts",
+    #         "Qb_topmodel": "sloth_Qb_topmodel",
+    #         "Qv_topmodel": "sloth_Qv_topmodel",
+    #         "global_deficit": "sloth_global_deficit",
+    #     },
+    #     overrides=[
+    #         (
+    #             "casam",
+    #             {
+    #                 "num_wetting_fronts": "soil_num_wetting_fronts",
+    #                 "soil_moisture_wetting_fronts": "soil_moisture_wetting_fronts",
+    #                 "soil_depth_wetting_fronts": "soil_depth_wetting_fronts",
+    #                 "soil_storage": "soil_storage",
+    #             },
+    #         ),
+    #         (
+    #             "cfe",
+    #             {
+    #                 "soil_storage": "SOIL_STORAGE",
+    #                 "soil_storage_change": "SOIL_STORAGE_CHANGE",
+    #             },
+    #         ),
+    #         (
+    #             "topmodel",
+    #             {
+    #                 "Qb_topmodel": "land_surface_water__baseflow_volume_flux",
+    #                 "Qv_topmodel": "soil_water_root-zone_unsat-zone_top__recharge_volume_flux",
+    #                 "global_deficit": "soil_water__domain_volume_deficit",
+    #             },
+    #         ),
+    #     ],
+    # ),
+    # "topmodel": ModelSpec(
+    #     main_output_variable="Qout",
+    #     realization_fragment=None,
+    #     routable=True,
+    #     variables_names_map={
+    #         "atmosphere_water__liquid_equivalent_precipitation_rate": "APCP_surface",
+    #         "water_potential_evaporation_flux": "sloth_pet",
+    #     },
+    #     overrides=[
+    #         (
+    #             "nom",
+    #             {
+    #                 "atmosphere_water__liquid_equivalent_precipitation_rate": "QINSUR",
+    #                 "water_potential_evaporation_flux": "EVAPOTRANS",
+    #             },
+    #         ),
+    #         ("snow17", {"atmosphere_water__liquid_equivalent_precipitation_rate": "raim"}),
+    #         ("pet", {"water_potential_evaporation_flux": "water_potential_evaporation_flux"}),
+    #     ],
+    # ),
+}
 
-    for _, row in divide_conf_df.iterrows():
-        with open(cat_config_dir / f"{row['divide_id']}.input", "w") as file:
-            file.write(
-                template.format(
-                    start_datetime=start_datetime,
-                    end_datetime=end_datetime,
-                    lat=row["latitude"],
-                    lon=row["longitude"],
-                    terrain_slope=row["mean.slope_1km"],
-                    azimuth=row["circ_mean.aspect"],
-                    ISLTYP=int(row["mode.ISLTYP"]),  # type: ignore
-                    IVGTYP=int(row["mode.IVGTYP"]),  # type: ignore
-                )
-            )
+# ---------------------------------------------------------------------------
+# ALL_SLOTH_MODEL_PARAMS
+# SLoTH model_params used to bootstrap variables at t=0 when upstream
+# models are absent. Format: { param_name: "(size,type,unit,location)" }
+# ---------------------------------------------------------------------------
+
+ALL_SLOTH_MODEL_PARAMS = {
+    "sloth_ice_fraction_schaake": "(1,double,m,node)",
+    "sloth_ice_fraction_xinanjiang": "(1,double,1,node)",
+    "sloth_soil_moisture_profile": "(1,double,1,node)",
+    "sloth_soil_temperature_profile": "(1,double,K,node)",
+    "sloth_soil_storage": "(1,double,m,node)",
+    "sloth_soil_storage_change": "(1,double,m,node)",
+    "sloth_num_wetting_fronts": "(1,double,1,node)",
+    "sloth_soil_moisture_wetting_fronts": "(1,double,1,node)",
+    "sloth_soil_depth_wetting_fronts": "(1,double,m,node)",
+    "sloth_Qb_topmodel": "(1,double,m h^-1,node)",
+    "sloth_Qv_topmodel": "(1,double,m h^-1,node)",
+    "sloth_global_deficit": "(1,double,m,node)",
+    "sloth_pet": "(1,double,m s-1,node)",
+    "sloth_ground_temperature": "(1,double,K,node)",
+}
+
+# ---------------------------------------------------------------------------
+# MODEL DEPENDENCY RULES
+# Format: (model, violation_lambda, message)
+# Lambda receives the full module list and returns True when VIOLATED.
+#
+# Read as: ("target_model", lambda models: <condition that means rule is broken>, "warning")
+# ---------------------------------------------------------------------------
+
+def _is_coupled(models: list[str]) -> bool:
+    if len([m for m in models if m != "sloth"]) > 1:
+        return True
+    return False
+
+MODEL_DEPENDENCY_RULES = (
+    # SLoTH required — bootstraps defaults for any model needing inter-model vars at t=0
+    ("cfe", lambda models: "sloth" not in models, "CFE requires SLoTH"),
+    ("casam", lambda models: "sloth" not in models, "CASAM requires SLoTH"),
+    # ("sft", lambda models: "sloth" not in models, "SFT requires SLoTH"),
+    # ("smp", lambda models: "sloth" not in models, "SMP requires SLoTH"),
+    # (
+    #     "smp",
+    #     lambda models: _is_coupled(models)
+    #     and not any(m in models for m in ("casam", "cfe", "topmodel")),
+    #     "SMP requires one of: CASAM, CFE, or TOPMODEL when coupled",
+    # ),
+    # (
+    #     "topmodel",
+    #     lambda models: "sloth" not in models and "pet" not in models and "nom" not in models,
+    #     "TOPMODEL requires SLoTH, NOM, or PET",
+    # ),
+    # (
+    #     "sac-sma",
+    #     lambda models: "sloth" not in models and "pet" not in models and "nom" not in models,
+    #     "SAC-SMA requires SLoTH, NOM, or PET",
+    # ),
+    # Snow17 must have a downstream runoff model (unless standalone with only SLoTH)
+    (
+        "snow17",
+        lambda models: _is_coupled(models)
+        and not any(m in models for m in ("cfe", "casam", "topmodel", "sac-sma")),
+        "Snow17 requires a downstream runoff model (CFE, CASAM, TOPMODEL, or SAC-SMA)",
+    ),
+    # CFE: NOM and Snow17 are mutually exclusive precip sources
+    (
+        "cfe",
+        lambda models: "nom" in models and "snow17" in models,
+        "CFE cannot have both NOM and Snow17 as precip sources",
+    ),
+    # NOM with no runoff model is likely misconfigured (unless standalone)
+    (
+        "nom",
+        lambda models: _is_coupled(models)
+        and not any(m in models for m in ("cfe", "casam", "topmodel", "sac-sma")),
+        "NOM is present but no runoff model (CFE, CASAM, TOPMODEL, SAC-SMA) found",
+    ),
+    # SFT: flag if coupled but no downstream consumer and no SMP
+    # (
+    #     "sft",
+    #     lambda models: _is_coupled(models)
+    #     and not any(m in models for m in ("cfe", "casam", "smp")),
+    #     "SFT has no downstream consumer (CFE or CASAM) and no SMP",
+    # ),
+    # Standalone models should not couple with physics models
+    (
+        "lstm",
+        lambda models: any(
+            m in models for m in ("cfe", "casam", "sft", "smp", "sac-sma", "topmodel")
+        ),
+        "LSTM is standalone — unexpected coupling with physics models",
+    ),
+    (
+        "lstm_rust",
+        lambda models: any(
+            m in models for m in ("cfe", "casam", "sft", "smp", "sac-sma", "topmodel")
+        ),
+        "LSTM-rust is standalone — unexpected coupling with physics models",
+    ),
+    (
+        "dhbv2",
+        lambda models: any(
+            m in models for m in ("cfe", "casam", "sft", "smp", "sac-sma", "topmodel")
+        ),
+        "dHBV2 is standalone — unexpected coupling with physics models",
+    ),
+    (
+        "dhbv2_daily",
+        lambda models: any(
+            m in models for m in ("cfe", "casam", "sft", "smp", "sac-sma", "topmodel")
+        ),
+        "dHBV2-daily is standalone — unexpected coupling with physics models",
+    ),
+    (
+        "summa",
+        lambda models: any(
+            m in models for m in ("cfe", "casam", "sft", "smp", "snow17", "pet", "sac-sma")
+        ),
+        "SUMMA is a full land surface model — unexpected coupling with physics models",
+    ),
+)
 
 
-def make_snow17_config(
-    base_dir: Path, divide_conf_df: pandas.DataFrame, start_time: datetime, end_time: datetime
-) -> None:
-    snow17_atts = duckdb.sql(f"""
-        SELECT * FROM '{FilePaths.snow17_attributes}'
-        WHERE divide_id IN {tuple(divide_conf_df["divide_id"])}
-    """).df()
+# This function would get called to use the above rules to validate a passed list of models
+def validate_models(models: list[str], routing: bool) -> list:
+    """Check that the specified models are valid and that any dependencies are met. If there are any
+    issues, return a list of warnings.
 
-    merged = snow17_atts.merge(
-        divide_conf_df[["divide_id", "areasqkm", "lengthkm", "latitude", "mean.elevation"]],
-        on="divide_id",
-    )
-
-    start_datetime = start_time.strftime("%Y%m%d%H")
-    end_datetime = end_time.strftime("%Y%m%d%H")
-    with open(FilePaths.template_snow17_config, "r") as config_file:
-        config_template = config_file.read()
-
-    cat_config_dir = base_dir / "cat_config" / "SNOW17"
-    cat_config_dir.mkdir(parents=True, exist_ok=True)
-
-    for _, row in merged.iterrows():
-        with open(cat_config_dir / f"{row['divide_id']}.input", "w") as file:
-            file.write(
-                config_template.format(
-                    divide_id=row["divide_id"],
-                    start_datetime=start_datetime,
-                    end_datetime=end_datetime,
-                )
-            )
-
-    with open(FilePaths.template_snow17_params, "r") as params_file:
-        params_template = params_file.read()
-
-    for _, row in merged.iterrows():
-        with open(cat_config_dir / f"params-{row['divide_id']}.txt", "w") as file:
-            file.write(
-                params_template.format(
-                    divide_id=row["divide_id"],
-                    areasqkm=row["areasqkm"],
-                    latitude=row["latitude"],
-                    elevation=row["mean.elevation"],
-                    scf=row["scf"],
-                    mfmax=row["mfmax"],
-                    mfmin=row["mfmin"],
-                    uadj=row["uadj"],
-                    si=row["si"],
-                    pxtemp=row["pxtemp"],
-                    nmf=row["nmf"],
-                    tipm=row["tipm"],
-                    mbase=row["mbase"],
-                    plwhc=row["plwhc"],
-                    daygm=row["daygm"],
-                    adc1=row["adc1"],
-                    adc2=row["adc2"],
-                    adc3=row["adc3"],
-                    adc4=row["adc4"],
-                    adc5=row["adc5"],
-                    adc6=row["adc6"],
-                    adc7=row["adc7"],
-                    adc8=row["adc8"],
-                    adc9=row["adc9"],
-                    adc10=row["adc10"],
-                    adc11=row["adc11"],
-                )
-            )
-
-
-def make_sacsma_config(
-    base_dir: Path, divide_conf_df: pandas.DataFrame, start_time: datetime, end_time: datetime
-) -> None:
-    sacsma_atts = duckdb.sql(f"""
-        SELECT * FROM '{FilePaths.sacsma_attributes}'
-        WHERE divide_id IN {tuple(divide_conf_df["divide_id"])}
-    """).df()
-
-    merged = sacsma_atts.merge(divide_conf_df[["divide_id", "areasqkm"]], on="divide_id")
-
-    start_datetime = start_time.strftime("%Y%m%d%H")
-    end_datetime = end_time.strftime("%Y%m%d%H")
-    with open(FilePaths.template_sac_config, "r") as config_file:
-        config_template = config_file.read()
-
-    cat_config_dir = base_dir / "cat_config" / "SAC-SMA"
-    cat_config_dir.mkdir(parents=True, exist_ok=True)
-
-    for _, row in merged.iterrows():
-        with open(cat_config_dir / f"{row['divide_id']}.input", "w") as file:
-            file.write(
-                config_template.format(
-                    divide_id=row["divide_id"],
-                    start_datetime=start_datetime,
-                    end_datetime=end_datetime,
-                )
-            )
-
-    with open(FilePaths.template_sac_params, "r") as params_file:
-        params_template = params_file.read()
-
-    for _, row in merged.iterrows():
-        with open(cat_config_dir / f"params-{row['divide_id']}.txt", "w") as file:
-            file.write(
-                params_template.format(
-                    divide_id=row["divide_id"],
-                    areasqkm=row["areasqkm"],
-                    uztwm=row["uztwm"],
-                    uzfwm=row["uzfwm"],
-                    lztwm=row["lztwm"],
-                    lzfpm=row["lzfpm"],
-                    lzfsm=row["lzfsm"],
-                    adimp=row["adimp"],
-                    uzk=row["uzk"],
-                    lzpk=row["lzpk"],
-                    lzsk=row["lzsk"],
-                    zperc=row["zperc"],
-                    rexp=row["rexp"],
-                    pctim=row["pctim"],
-                    pfree=row["pfree"],
-                    riva=row["riva"],
-                    side=row["side"],
-                    rserv=row["rserv"],
-                )
-            )
-
-
-def get_model_attributes(hydrofabric: Path, layer: str = "divides") -> pandas.DataFrame:
-    with sqlite3.connect(hydrofabric) as conn:
-        conf_df = pandas.read_sql_query(
-            """
-            SELECT
-            d.areasqkm,
-            d.lengthkm,
-            da.*
-            FROM divides AS d
-            JOIN 'divide-attributes' AS da ON d.divide_id = da.divide_id
-            """,
-            conn,
-        )
-    source_crs = get_table_crs_short(hydrofabric, layer)
-    transformer = Transformer.from_crs(source_crs, "EPSG:4326", always_xy=True)
-    lon, lat = transformer.transform(conf_df["centroid_x"].values, conf_df["centroid_y"].values)
-    conf_df["longitude"] = lon
-    conf_df["latitude"] = lat
-    return conf_df
-
-
-def make_lstm_config(
-    hydrofabric: Path,
-    output_dir: Path,
-    template_path: Path = FilePaths.template_lstm_config,
-):
-    divide_conf_df = get_model_attributes(hydrofabric)
-
-    cat_config_dir = output_dir / "cat_config" / "lstm"
-    if cat_config_dir.exists():
-        shutil.rmtree(cat_config_dir)
-    cat_config_dir.mkdir(parents=True, exist_ok=True)
-
-    # convert the mean.slope from degrees 0-90 where 90 is flat and 0 is vertical to m/km
-    # flip 0 and 90 degree values
-    divide_conf_df["flipped_mean_slope"] = abs(divide_conf_df["mean.slope"] - 90)
-    # Convert degrees to meters per kmmeter
-    divide_conf_df["mean_slope_mpkm"] = (
-        np.tan(np.radians(divide_conf_df["flipped_mean_slope"])) * 1000
-    )
-
-    with open(template_path, "r") as file:
-        template = file.read()
-
-    for _, row in divide_conf_df.iterrows():
-        divide = row["divide_id"]
-        with open(cat_config_dir / f"{divide}.yml", "w") as file:
-            file.write(
-                template.format(
-                    area_sqkm=row["areasqkm"],
-                    divide_id=divide,
-                    lat=row["latitude"],
-                    lon=row["longitude"],
-                    slope_mean=row["mean_slope_mpkm"],
-                    elevation_mean=row["mean.elevation"] / 100,  # convert cm in hf to m
-                )
-            )
-
-
-def make_dhbv2_config(
-    hydrofabric: Path,
-    output_dir: Path,
-    start_time: datetime,
-    end_time: datetime,
-    template_path: Path = FilePaths.template_dhbv2_config,
-    output_suffix: str = "",
-    clear_dir: bool = True,
-):
-    divide_conf_df = get_model_attributes(hydrofabric)
-    dhbv_atts = duckdb.sql(f"""
-        SELECT * FROM '{FilePaths.dhbv_attributes}'
-        WHERE divide_id IN {tuple(divide_conf_df["divide_id"])}
-    """).df()
-    if template_path == FilePaths.template_dhbv2_daily_config:
-        cat_config_dir = output_dir / "cat_config" / "dhbv2_daily"
-    else:
-        cat_config_dir = output_dir / "cat_config" / "dhbv2"
-    if cat_config_dir.exists() and clear_dir:
-        shutil.rmtree(cat_config_dir)
-    cat_config_dir.mkdir(parents=True, exist_ok=True)
-
-    template = template_path.read_text()
-    merged = dhbv_atts.merge(
-        divide_conf_df[["divide_id", "areasqkm", "lengthkm", "latitude"]], on="divide_id"
-    )
-
-    for _, row in merged.iterrows():
-        (cat_config_dir / f"{row['divide_id']}{output_suffix}.yml").write_text(
-            template.format(
-                **row,
-                start_time=start_time,
-                end_time=end_time,
-                start_date=start_time.strftime("%Y/%m/%d"),
-                start_time_str=start_time.strftime("%Y/%m/%d %H"),
-            )
-        )
-
-
-def make_summa_config(hru_ids: list[int], output_dir: Path):
-    with open(FilePaths.template_summa_config, "r") as file:
-        template = file.read()
-    cat_config_dir = output_dir / "cat_config" / "SUMMA"
-    cat_config_dir.mkdir(parents=True, exist_ok=True)
-    for i, id in enumerate(hru_ids):
-        divide = f"cat-{id}"
-        with open(cat_config_dir / f"{divide}.input", "w") as file:
-            # + 1 because fortran's first index is 1 not 0
-            file.write(template.format(divide_index=i + 1))
-
-def make_casam_config(
-    base_dir: Path,
-    divide_conf_df: pandas.DataFrame,
-    start_time: datetime,
-    end_time: datetime,
-    sft_coupled: bool
-) -> None:
-    """Generates CASAM BMI config
+    models[-1] determines main_output_variable, which is the variable routed through t-route. So
+    if models[-1] is not a rainfall-runoff model, a warning will be returned.
 
     Args:
-        base_dir (Path): ngen run directory, usually the name of the cat-ID
-        divide_conf_df (pandas.DataFrame): Subsetted hydrofabric layer
-        start_time (datetime): Start time of simulation
-        end_time (datetime): End time of simulation
-        sft_coupled (bool): True if SFT is coupled, False if not
+        models (list[str]): List of models to use, in the order they will be executed
+        routing (bool): Whether routing is enabled
+
+    Returns:
+        warnings (lsit[str]): List of warnings about unmet dependencies
     """
-    casam_atts = duckdb.sql(f"""
-        SELECT * FROM '{FilePaths.casam_attributes}'
-        WHERE divide_id IN {tuple(divide_conf_df["divide_id"])}
-    """).df()
 
-    def fmt_list(value) -> str:
-        return ",".join(str(v) for v in value)
+    if len(models) == 0:
+        raise ValueError("No models specified")
 
-    elapsed_time = end_time - start_time
-    elapsed_hours = elapsed_time.total_seconds() / 3600
-
-    with open(FilePaths.template_casam_config, "r", encoding="utf-8") as config_file:
-        config_template = config_file.read()
-
-    cat_config_dir = base_dir / "cat_config" / "CASAM"
-    cat_config_dir.mkdir(parents=True, exist_ok=True)
-
-    for _, row in casam_atts.iterrows():
-        with open(cat_config_dir / f"{row['divide_id']}.input", "w", encoding="utf-8") as file:
-            file.write(
-                config_template.format(
-                    layer_thickness=fmt_list(row["layer_thickness"]),
-                    initial_psi=row["initial_psi"],
-                    ponded_depth_max=row["ponded_depth_max"],
-                    endtime=elapsed_hours,
-                    layer_soil_type=fmt_list(row["layer_soil_type"]),
-                    max_valid_soil_types=row["max_valid_soil_types"],
-                    wilting_point_psi=row["wilting_point_psi"],
-                    field_capacity_psi=row["field_capacity_psi"],
-                    giuh_ordinates=fmt_list(row["giuh_ordinates"]),
-                    sft_coupled=str(sft_coupled).lower(),
-                    soil_z=fmt_list(row["soil_z"])
-                )
+    if any(model not in MODEL_REGISTRY for model in models):
+        invalid_models = [model for model in models if model not in MODEL_REGISTRY]
+        raise ValueError(
+            (
+                f"Invalid models specified: {invalid_models}. " +
+                f"Supported models are: {list(MODEL_REGISTRY.keys())}"
             )
+        )
+
+    # checks model dependencies
+    warnings = []
+    warnings.extend(
+        message
+        for model_name, predicate, message in MODEL_DEPENDENCY_RULES
+        if model_name in models and predicate(models)
+    )
+
+    # Check that a rainfall-runoff model is used as the final model when routing is on
+    if routing and not MODEL_REGISTRY[models[-1]].routable:
+        warnings.append("Routing is on but no rainfall-runoff model is used")
+
+    return warnings
 
 
-def configure_troute(
-    cat_id: str, config_dir: Path, start_time: datetime, end_time: datetime
+def _insert_sloth_module(
+    models: list[str], target_variable_names: dict[str, dict[str, str]], modules: list[dict]
 ) -> None:
-    with open(FilePaths.template_troute_config, "r") as file:
-        troute_template = file.read()
-    time_step_size = 300
-    gpkg_file_path = f"{config_dir}/{cat_id}_subset.gpkg"
-    nts = (end_time - start_time).total_seconds() / time_step_size
-    with sqlite3.connect(gpkg_file_path) as conn:
-        ncats_df = pandas.read_sql_query("SELECT COUNT(id) FROM 'divides';", conn)
-        ncats = ncats_df["COUNT(id)"][0]
+    params: dict[str, float] = {}
+    for model_vars in target_variable_names.values():
+        for varname in model_vars.values():
+            if varname in ALL_SLOTH_MODEL_PARAMS:
+                params[varname + ALL_SLOTH_MODEL_PARAMS[varname]] = 0.0
+    sloth_position = models.index("sloth")
+    with open(
+        MODEL_REGISTRY["sloth"].realization_fragment, "r", encoding="utf-8" # type: ignore
+    ) as f:
+        sloth_realization = json.load(f)
 
-    est_bytes_required = nts * ncats * 45  # extremely rough calculation based on about 3 tests :)
-    local_ram_available = (
-        0.8 * psutil.virtual_memory().available
-    )  # buffer to not accidentally explode machine
+    # insert dummy param if no other params are present, otherwise BMI will throw an error
+    if not params:
+        params["sloth_dummy_param(1,double,1,node)"] = 0.0
+    sloth_realization["params"]["model_params"] = params
+    modules.insert(sloth_position, sloth_realization)
 
-    if est_bytes_required > local_ram_available:
-        max_loop_size = nts // (est_bytes_required // local_ram_available)
-        binary_nexus_file_folder_comment = ""
-        parent_dir = config_dir.parent
-        output_parquet_path = Path(f"{parent_dir}/outputs/parquet/")
 
-        if not output_parquet_path.exists():
-            os.makedirs(output_parquet_path)
-    else:
-        max_loop_size = nts
-        binary_nexus_file_folder_comment = "#"
-
-    filled_template = troute_template.format(
-        # hard coded to 5 minutes
-        time_step_size=time_step_size,
-        # troute seems to be ok with setting this to your cpu_count
-        cpu_pool=multiprocessing.cpu_count(),
-        geo_file_path=f"./config/{cat_id}_subset.gpkg",
-        start_datetime=start_time.strftime("%Y-%m-%d %H:%M:%S"),
-        nts=nts,
-        max_loop_size=max_loop_size,
-        binary_nexus_file_folder_comment=binary_nexus_file_folder_comment,
+def _handle_calibrated_params(paths: FilePaths, gage_id: str) -> bool:
+    # try and download s3:communityhydrofabric/hydrofabrics/community/gage_parameters/gage_id
+    # if it doesn't exist, use the default
+    url = (
+        "https://communityhydrofabric.s3.us-east-1.amazonaws.com/hydrofabrics/community/"
+        + f"gage_parameters/{gage_id}.json"
     )
+    response = requests.get(url, timeout=10)
 
-    with open(config_dir / "troute.yaml", "w") as file:
-        file.write(filled_template)
+    if response.status_code == 200:
+        new_template = response.json()
+        template_path = paths.config_dir / "downloaded_params.json"
+        with open(template_path, "w", encoding="utf-8") as f:
+            json.dump(new_template, f)
+
+        logger.info("downloaded calibrated parameters for %s", gage_id)
+        return True
+
+    logger.warning("could not download parameters for %s, using default template", gage_id)
+    return False
 
 
-def make_ngen_realization_json(
-    config_dir: Path,
-    template_path: Path,
+def create_modular_realization(  # pylint: disable=too-many-locals,too-many-arguments
+    output_folder: str,
     start_time: datetime,
     end_time: datetime,
-    output_interval: int = 3600,
-) -> None:
-    with open(template_path, "r") as file:
-        realization = json.load(file)
-
-    realization["time"]["start_time"] = start_time.strftime("%Y-%m-%d %H:%M:%S")
-    realization["time"]["end_time"] = end_time.strftime("%Y-%m-%d %H:%M:%S")
-    realization["time"]["output_interval"] = output_interval
-
-    with open(config_dir / "realization.json", "w") as file:
-        json.dump(realization, file, indent=4)
-
-
-def create_lstm_realization(
-    cat_id: str, start_time: datetime, end_time: datetime, use_rust: bool = False
+    models: list[str],
+    *,
+    routing: bool = False,
+    gage_id: str | None = None,
 ):
-    paths = FilePaths(cat_id)
-    realization_path = paths.config_dir / "realization.json"
-    configure_troute(cat_id, paths.config_dir, start_time, end_time)
-    # python version of the lstm
-    python_template_path = FilePaths.template_lstm_realization_config
-    make_ngen_realization_json(paths.config_dir, python_template_path, start_time, end_time)
-    realization_path.rename(paths.config_dir / "python_lstm_real.json")
-    # rust version of the lstm
-    rust_template_path = FilePaths.template_lstm_rust_realization_config
-    make_ngen_realization_json(paths.config_dir, rust_template_path, start_time, end_time)
-    realization_path.rename(paths.config_dir / "rust_lstm_real.json")
+    """Creates a realization file based on the specified models.
+    Note: This function automatically fetches calibrated parameters. Calibrated params are only
+    available at certain gages for the SLoTH, NOM, and CFE model combination.
 
-    if use_rust:
-        (paths.config_dir / "rust_lstm_real.json").rename(realization_path)
-    else:
-        (paths.config_dir / "python_lstm_real.json").rename(realization_path)
+    models[-1] determines main_output_variable, which is the variable routed through t-route.
 
-    make_lstm_config(paths.geopackage_path, paths.config_dir)
-    paths.setup_run_folders()
+    Args:
+        output_folder (str): Name of the output folder, usually the cat-id
+        start_time (datetime): Start time of simulation in YYYY-MM-DD HH:MM:SS
+        end_time (datetime): End time of simulation in YYYY-MM-DD HH:MM:SS
+        models (list[str]): List of models to be coupled together, in the order that they will be
+            executed
+        routing (bool, optional): True if t-route is coupled. Defaults to False.
+        gage_id (str | None, optional): Gage ID for the simulation. Defaults to None.
+    """
 
+    paths = FilePaths(output_folder)
 
-def create_dhbv2_realization(
-    cat_id: str, start_time: datetime, end_time: datetime, daily: bool = False
-):
-    paths = FilePaths(cat_id)
-    configure_troute(cat_id, paths.config_dir, start_time, end_time)
+    # calibrated parameter fetching
+    if gage_id is not None and models == ["sloth", "nom", "cfe"]:
+        if _handle_calibrated_params(paths, gage_id):
+            return
 
-    if daily:
-        make_ngen_realization_json(
-            paths.config_dir,
-            FilePaths.template_dhbv2_daily_realization_config,
-            start_time,
-            end_time,
-        )
-        make_dhbv2_config(
-            paths.geopackage_path,
-            paths.config_dir,
-            start_time,
-            end_time,
-            template_path=FilePaths.template_dhbv2_daily_config,
-        )
-    else:
-        make_ngen_realization_json(
-            paths.config_dir,
-            FilePaths.template_dhbv2_realization_config,
-            start_time,
-            end_time,
-        )
-        make_dhbv2_config(paths.geopackage_path, paths.config_dir, start_time, end_time)
+    main_output_variable = MODEL_REGISTRY[models[-1]].main_output_variable
 
-    paths.setup_run_folders()
+    target_variable_names = {}
+    for model in models:
+        if MODEL_REGISTRY[model].variables_names_map:
+            target_variable_names[model] = copy.deepcopy(MODEL_REGISTRY[model].variables_names_map)
 
+    modules: list[dict] = []
+    seen_models: list[str] = []
 
-def get_hru_order(forcing_path: Path) -> list[int]:
-    # the SUMMA hru (hydrologic response unit) is like a nextgen hydrofabric catchment
-    # to correctly format the input data we need the order of these ids to be consistent
-    if not forcing_path.exists():
-        raise FileNotFoundError(
-            f"Unable to create SUMMA configuration without forcing file {forcing_path}"
-        )
-    forcings = xr.open_dataset(forcing_path)
-    return [int(s.split("-")[-1]) for s in forcings.ids.values]
+    for model in models:
+        if model == "sloth":
+            continue  # SLoTH is handled separately below after all other models have been processed
 
+        # Implicitly this means that if we have something like ["nom", "pet"], then the PET value
+        # from the evapotranspiration module will override the PET value from Noah-OWP-M
+        for dependency, overrides in MODEL_REGISTRY[model].overrides:
+            if dependency in seen_models:
+                target_variable_names[model].update(overrides)
 
-def make_summa_attributes(hru_ids, hydrofabric):
-    cat_ids = [f"cat-{id}" for id in hru_ids]
-    divide_conf_df = get_model_attributes(hydrofabric)
-    divide_conf_df = divide_conf_df.set_index("divide_id")
+        with open(
+            MODEL_REGISTRY[model].realization_fragment, "r", encoding="utf-8" # type: ignore
+        ) as f:
+            realization = json.load(f)
+        if model in target_variable_names:
+            realization["params"]["variables_names_map"] = target_variable_names[model]
+        modules.append(realization)
 
-    # Validate all IDs exist
-    for hid in cat_ids:
-        if hid not in divide_conf_df.index:
-            raise ValueError(f"HRU ID {hid} is not present in hydrofabric model attributes")
+        seen_models.append(model)
 
-    # Subset and preserve order
-    df = divide_conf_df.loc[cat_ids]
+    if "sloth" in models:
+        _insert_sloth_module(models, target_variable_names, modules)
 
-    n_hru = len(hru_ids)
-    hru_id_array = np.array(hru_ids, dtype=np.int32)
+    with open(FilePaths.modular_template, "r", encoding="utf-8") as f:
+        realization = json.load(f)
 
-    # Convert area from sq km to sq m
-    hru_area = df["areasqkm"].values * 1e6
+    realization["global"]["formulations"][0]["params"][
+        "main_output_variable"
+    ] = main_output_variable
+    realization["global"]["formulations"][0]["params"]["modules"] = modules
+    realization["time"]["start_time"] = datetime.strftime(start_time, "%Y-%m-%d %H:%M:%S")
+    realization["time"]["end_time"] = datetime.strftime(end_time, "%Y-%m-%d %H:%M:%S")
 
-    # Convert elevation from cm to m
-    elevation = df["mean.elevation"].values / 100.0
+    if routing:
+        realization["routing"] = {"t_route_config_file_with_path": "./config/troute.yaml"}
 
-    # Convert mean.slope (degrees, 90=flat 0=vertical) to tan_slope (m/m)
-    flipped = np.abs(df["mean.slope"].values - 90)
-    tan_slope = np.tan(np.radians(flipped))
-
-    ds = xr.Dataset(
-        {
-            "hruId": xr.DataArray(
-                data=hru_id_array,
-                dims=["hru"],
-                attrs={"units": "-", "long_name": "Index of hydrological response unit (HRU)"},
-            ),
-            "gruId": xr.DataArray(
-                data=hru_id_array.copy(),
-                dims=["gru"],
-                attrs={"units": "-", "long_name": "Index of grouped response unit (GRU)"},
-            ),
-            "hru2gruId": xr.DataArray(
-                data=hru_id_array.copy(),
-                dims=["hru"],
-                attrs={"units": "-", "long_name": "Index of GRU to which the HRU belongs"},
-            ),
-            "downHRUindex": xr.DataArray(
-                data=np.full(n_hru, 0, dtype=np.int32),
-                dims=["hru"],
-                attrs={"units": "-", "long_name": "Index of downslope HRU (0 = basin outlet)"},
-            ),
-            "longitude": xr.DataArray(
-                data=df["longitude"].values.astype(np.float64),
-                dims=["hru"],
-                attrs={"units": "Decimal degree east", "long_name": "Longitude of HRUs centroid"},
-            ),
-            "latitude": xr.DataArray(
-                data=df["latitude"].values.astype(np.float64),
-                dims=["hru"],
-                attrs={"units": "Decimal degree north", "long_name": "Latitude of HRUs centroid"},
-            ),
-            "elevation": xr.DataArray(
-                data=elevation.astype(np.float64),
-                dims=["hru"],
-                attrs={"units": "m", "long_name": "Mean HRU elevation"},
-            ),
-            "HRUarea": xr.DataArray(
-                data=hru_area.astype(np.float64),
-                dims=["hru"],
-                attrs={"units": "m^2", "long_name": "Area of HRU"},
-            ),
-            "tan_slope": xr.DataArray(
-                data=tan_slope.astype(np.float64),
-                dims=["hru"],
-                attrs={"units": "m m-1", "long_name": "Average tangent slope of HRU"},
-            ),
-            "contourLength": xr.DataArray(
-                data=np.full(n_hru, 100.0, dtype=np.float64),
-                dims=["hru"],
-                attrs={"units": "m", "long_name": "Contour length of HRU"},
-            ),
-            "slopeTypeIndex": xr.DataArray(
-                data=np.full(n_hru, 1, dtype=np.int32),
-                dims=["hru"],
-                attrs={"units": "-", "long_name": "Index defining slope"},
-            ),
-            "soilTypeIndex": xr.DataArray(
-                data=df["mode.ISLTYP"].values.astype(np.int32),
-                dims=["hru"],
-                attrs={"units": "-", "long_name": "Index defining soil type"},
-            ),
-            "vegTypeIndex": xr.DataArray(
-                data=df["mode.IVGTYP"].values.astype(np.int32),
-                dims=["hru"],
-                attrs={"units": "-", "long_name": "Index defining vegetation type"},
-            ),
-            "mHeight": xr.DataArray(
-                data=np.full(n_hru, 1.5, dtype=np.float64),
-                dims=["hru"],
-                attrs={"units": "m", "long_name": "Measurement height above bare ground"},
-            ),
-        },
-        attrs={
-            "Author": "Created by ngiab preprocessor SUMMA workflow script",
-            "History": f"Created {datetime.now().strftime('%Y/%m/%d %H:%M:%S')}",
-        },
-    )
-
-    return ds
-
-
-def make_summa_trialParams(hru_ids: list[int], timesteps: int) -> xr.Dataset:
-    ds = xr.Dataset(
-        {
-            "hruId": xr.DataArray(
-                data=np.array(hru_ids, dtype=np.int32),
-                dims=["hru"],
-                attrs={"units": "-", "long_name": "Index of hydrological response unit (HRU)"},
-            ),
-            "maxstep": xr.DataArray(
-                data=np.full(len(hru_ids), timesteps * 3600, dtype=np.float64),
-                dims=["hru"],
-            ),
-        },
-        attrs={
-            "Author": "Created by ngiab preprocessor SUMMA workflow script",
-            "History": f"Created {datetime.now().strftime('%Y/%m/%d %H:%M:%S')}",
-            "Purpose": "Create a trial parameter .nc file for initial SUMMA runs",
-        },
-    )
-    return ds
-
-
-def make_summa_coldState(hru_ids):
-    n_hru = len(hru_ids)
-    n_midToto = 3
-    n_ifcToto = 4
-
-    def scalar_var(fill_val, dtype=np.float64):
-        return xr.DataArray(
-            data=np.full((1, n_hru), fill_val, dtype=dtype),
-            dims=["scalarv", "hru"],
-        )
-
-    def layer_var(fill_val, dim_name, dim_size):
-        return xr.DataArray(
-            data=np.full((dim_size, n_hru), fill_val, dtype=np.float64),
-            dims=[dim_name, "hru"],
-        )
-
-    iLayerHeight_data = np.broadcast_to(
-        np.array([0.0, 0.2, 0.5, 1.0])[:, np.newaxis],
-        (n_ifcToto, n_hru),
-    ).copy()
-
-    mLayerDepth_data = np.broadcast_to(
-        np.array([0.2, 0.3, 0.5])[:, np.newaxis],
-        (n_midToto, n_hru),
-    ).copy()
-
-    ds = xr.Dataset(
-        {
-            "hruId": xr.DataArray(
-                data=np.array(hru_ids, dtype=np.int32),
-                dims=["hru"],
-                attrs={"units": "-", "long_name": "Index of hydrological response unit (HRU)"},
-            ),
-            "dt_init": scalar_var(3600.0),
-            "nSoil": scalar_var(3, dtype=np.int32),
-            "nSnow": scalar_var(0, dtype=np.int32),
-            "scalarCanopyIce": scalar_var(0.0),
-            "scalarCanopyLiq": scalar_var(0.0),
-            "scalarSnowDepth": scalar_var(0.0),
-            "scalarSWE": scalar_var(0.0),
-            "scalarSfcMeltPond": scalar_var(0.0),
-            "scalarAquiferStorage": scalar_var(2.5),
-            "scalarSnowAlbedo": scalar_var(0.0),
-            "scalarCanairTemp": scalar_var(283.16),
-            "scalarCanopyTemp": scalar_var(283.16),
-            "mLayerTemp": layer_var(283.16, "midToto", n_midToto),
-            "mLayerVolFracIce": layer_var(0.0, "midToto", n_midToto),
-            "mLayerVolFracLiq": layer_var(0.2, "midToto", n_midToto),
-            "mLayerMatricHead": layer_var(-1.0, "midToto", n_midToto),
-            "iLayerHeight": xr.DataArray(data=iLayerHeight_data, dims=["ifcToto", "hru"]),
-            "mLayerDepth": xr.DataArray(data=mLayerDepth_data, dims=["midToto", "hru"]),
-        },
-        attrs={
-            "Author": "Created by SUMMA workflow scripts",
-            "History": f"Created {datetime.now().strftime('%Y/%m/%d %H:%M:%S')}",
-            "Purpose": "Create a cold state .nc file for initial SUMMA runs",
-        },
-    )
-
-    # Add midSoil dimension (SUMMA expects it even though no variable uses it)
-    ds = ds.assign_coords(midSoil=np.arange(n_midToto))
-
-    # Build encoding to suppress _FillValue on all variables
-    encoding = {var: {"_FillValue": None} for var in ds.data_vars}
-
-    return ds, encoding
-
-
-def create_summa_realization(cat_id: str, start_time: datetime, end_time: datetime):
-    paths = FilePaths(cat_id)
-    configure_troute(cat_id, paths.config_dir, start_time, end_time)
-    make_ngen_realization_json(
-        paths.config_dir,
-        FilePaths.template_summa_realization_config,
-        start_time,
-        end_time,
-    )
-    paths.summa_model_config.mkdir(parents=True, exist_ok=True)
-
-    files = chain(
-        FilePaths.summa_file_dir.glob("*.txt"),
-        FilePaths.summa_file_dir.glob("*.TBL"),
-        FilePaths.summa_file_dir.glob("*.md"),
-    )
-    for file in files:
-        if file.name == "fileManager.txt":
-            with open(file, "r") as f:
-                template = f.read()
-            with open(paths.summa_model_config / file.name, "w") as f:
-                f.write(template.format(start_time=start_time, end_time=end_time))
-        else:
-            shutil.copy(file, paths.summa_model_config)
-
-    max_timesteps = int((end_time - start_time).total_seconds() / 3600)
-    hru_ids = get_hru_order(paths.forcings_file)
-    make_summa_attributes(hru_ids, FilePaths.conus_hydrofabric).to_netcdf(
-        paths.summa_model_config / "attributes.nc"
-    )
-    make_summa_trialParams(hru_ids, max_timesteps).to_netcdf(
-        paths.summa_model_config / "trialParams.nc"
-    )
-    ds, encoding = make_summa_coldState(hru_ids)
-    ds.to_netcdf(paths.summa_model_config / "coldState.nc", encoding=encoding)
-    make_summa_config(hru_ids, paths.config_dir)
-    paths.setup_run_folders(["outputs/summa"])
-
-
-def create_snow17_realization(
-    cat_id: str,
-    start_time: datetime,
-    end_time: datetime,
-    use_nwm_gw: bool = False,
-    gage_id: Optional[str] = None,
-):
-    paths = FilePaths(cat_id)
-
-    conf_df = get_model_attributes(paths.geopackage_path)
-    if use_nwm_gw:
-        gw_levels = get_approximate_gw_storage(paths, start_time)
-    else:
-        gw_levels = dict()
-
-    make_cfe_config(conf_df, paths, gw_levels)
-    make_noahowp_config(paths.config_dir, conf_df, start_time, end_time)
-    configure_troute(cat_id, paths.config_dir, start_time, end_time)
-    make_snow17_config(paths.config_dir, conf_df, start_time, end_time)
-
-    make_ngen_realization_json(
-        paths.config_dir,
-        FilePaths.template_snow17_realization_config,
-        start_time,
-        end_time,
-    )
-    paths.setup_run_folders()
-
-
-def create_sacsma_realization(
-    cat_id: str, start_time: datetime, end_time: datetime, gage_id: Optional[str] = None
-):
-    paths = FilePaths(cat_id)
-
-    conf_df = get_model_attributes(paths.geopackage_path)
-
-    make_noahowp_config(paths.config_dir, conf_df, start_time, end_time)
-    make_sacsma_config(paths.config_dir, conf_df, start_time, end_time)
-    configure_troute(cat_id, paths.config_dir, start_time, end_time)
-    make_ngen_realization_json(
-        paths.config_dir,
-        FilePaths.template_sac_realization_config,
-        start_time,
-        end_time,
-    )
-    paths.setup_run_folders()
-
-
-def create_realization(
-    cat_id: str,
-    start_time: datetime,
-    end_time: datetime,
-    use_nwm_gw: bool = False,
-    gage_id: Optional[str] = None,
-):
-    paths = FilePaths(cat_id)
-
-    template_path = paths.template_cfe_nowpm_realization_config
-
-    if gage_id is not None:
-        # try and download s3:communityhydrofabric/hydrofabrics/community/gage_parameters/gage_id
-        # if it doesn't exist, use the default
-        url = f"https://communityhydrofabric.s3.us-east-1.amazonaws.com/hydrofabrics/community/gage_parameters/{gage_id}.json"
-        response = requests.get(url)
-        if response.status_code == 200:
-            new_template = requests.get(url).json()
-            template_path = paths.config_dir / "downloaded_params.json"
-            with open(template_path, "w") as f:
-                json.dump(new_template, f)
-            logger.info(f"downloaded calibrated parameters for {gage_id}")
-        else:
-            logger.warning(f"could not download parameters for {gage_id}, using default template")
-
-    conf_df = get_model_attributes(paths.geopackage_path)
-
-    if use_nwm_gw:
-        gw_levels = get_approximate_gw_storage(paths, start_time)
-    else:
-        gw_levels = dict()
-
-    make_cfe_config(conf_df, paths, gw_levels)
-
-    make_noahowp_config(paths.config_dir, conf_df, start_time, end_time)
-
-    configure_troute(cat_id, paths.config_dir, start_time, end_time)
-
-    make_ngen_realization_json(paths.config_dir, template_path, start_time, end_time)
-
-    # create some partitions for parallelization
-    paths.setup_run_folders()
+    with open(paths.config_dir / "realization.json", "w", encoding="utf-8") as f:
+        json.dump(realization, f, indent=4)
