+++ modules/data_sources/config/modular_realization/cfe.json
@@ -10,11 +10,6 @@
         "uses_forcing_file": false,
         "registration_function": "register_bmi_cfe",
         "variables_names_map": {
-            "water_potential_evaporation_flux": "EVAPOTRANS",
-            "atmosphere_water__liquid_equivalent_precipitation_rate": "QINSUR",
-            "ice_fraction_schaake": "sloth_ice_fraction_schaake",
-            "ice_fraction_xinanjiang": "sloth_ice_fraction_xinanjiang",
-            "soil_moisture_profile": "sloth_soil_moisture_profile"
         },
         "library_file": "/dmod/shared_libs/libcfebmi.so.1.0.0"
     }
