+++ modules/data_sources/config/modular_realization/dhbv2_daily.json
@@ -9,15 +9,6 @@
         "fixed_time_step": true,
         "uses_forcing_file": false,
         "variables_names_map": {
-            "atmosphere_water__liquid_equivalent_precipitation_rate": "precip_rate",
-            "land_surface_air__temperature": "TMP_2maboveground",
-            "atmosphere_air_water~vapor__relative_saturation": "SPFH_2maboveground",
-            "land_surface_radiation~incoming~longwave__energy_flux": "DLWRF_surface",
-            "land_surface_radiation~incoming~shortwave__energy_flux": "DSWRF_surface",
-            "land_surface_air__pressure": "PRES_surface",
-            "land_surface_wind__x_component_of_velocity": "UGRD_10maboveground",
-            "land_surface_wind__y_component_of_velocity": "VGRD_10maboveground",
-            "land_surface_water__runoff_volume_flux": "streamflow"
         }
     }
 }
