+++ modules/data_sources/config/modular_realization/summa.json
@@ -8,14 +8,6 @@
         "allow_exceed_end_time": true,
         "main_output_variable": "land_surface_water__runoff_volume_flux",
         "variables_names_map": {
-            "atmosphere_water__precipitation_mass_flux": "precip_rate",
-            "land_surface_air__temperature": "TMP_2maboveground",
-            "atmosphere_air_water~vapor__relative_saturation": "SPFH_2maboveground",
-            "land_surface_wind__x_component_of_velocity": "UGRD_10maboveground",
-            "land_surface_wind__y_component_of_velocity": "VGRD_10maboveground",
-            "land_surface_radiation~incoming~shortwave__energy_flux": "DSWRF_surface",
-            "land_surface_radiation~incoming~longwave__energy_flux": "DLWRF_surface",
-            "land_surface_air__pressure": "PRES_surface"
         },
         "uses_forcing_file": false
     }
