--- modules/data_sources/config/realization/snow17-nom-cfe.json
@@ -1,114 +1,0 @@
-{
-    "global": {
-        "formulations": [
-            {
-                "name": "bmi_multi",
-                "params": {
-                    "name": "bmi_multi",
-                    "model_type_name": "bmi_multi",
-                    "main_output_variable": "Q_OUT",
-                    "forcing_file": "",
-                    "init_config": "",
-                    "allow_exceed_end_time": true,
-                    "modules": [
-                        {
-                            "name": "bmi_c++",
-                            "params": {
-                                "name": "bmi_c++",
-                                "model_type_name": "SLOTH",
-                                "main_output_variable": "z",
-                                "init_config": "/dev/null",
-                                "allow_exceed_end_time": true,
-                                "fixed_time_step": false,
-                                "uses_forcing_file": false,
-                                "model_params": {
-                                    "sloth_ice_fraction_schaake(1,double,m,node)": 0.0,
-                                    "sloth_ice_fraction_xinanjiang(1,double,1,node)": 0.0,
-                                    "sloth_soil_moisture_profile(1,double,1,node)": 0.0
-                                },
-                                "library_file": "/dmod/shared_libs/libslothmodel.so",
-                                "registration_function": "none"
-                            }
-                        },
-                        {
-                            "name": "bmi_fortran",
-                            "params": {
-                                "name": "bmi_fortran",
-                                "model_type_name": "SNOW17",
-                                "library_file": "/dmod/shared_libs/libsnow17bmi.so",
-                                "forcing_file": "",
-                                "init_config": "./config/cat_config/SNOW17/{{id}}.input",
-                                "allow_exceed_end_time": true,
-                                "main_output_variable": "raim",
-                                "variables_names_map": {
-                                    "precip": "atmosphere_water__liquid_equivalent_precipitation_rate",
-                                    "tair": "land_surface_air__temperature"
-                                },
-                                "uses_forcing_file": false
-                            }
-                        },
-                        {
-                            "name": "bmi_fortran",
-                            "params": {
-                                "name": "bmi_fortran",
-                                "model_type_name": "NoahOWP",
-                                "library_file": "/dmod/shared_libs/libsurfacebmi.so",
-                                "forcing_file": "",
-                                "init_config": "./config/cat_config/NOAH-OWP-M/{{id}}.input",
-                                "allow_exceed_end_time": true,
-                                "main_output_variable": "QINSUR",
-                                "variables_names_map": {
-                                    "PRCPNONC": "precip_rate",
-                                    "Q2": "SPFH_2maboveground",
-                                    "SFCTMP": "TMP_2maboveground",
-                                    "UU": "UGRD_10maboveground",
-                                    "VV": "VGRD_10maboveground",
-                                    "LWDN": "DLWRF_surface",
-                                    "SOLDN": "DSWRF_surface",
-                                    "SFCPRS": "PRES_surface"
-                                },
-                                "uses_forcing_file": false
-                            }
-                        },
-                        {
-                            "name": "bmi_c",
-                            "params": {
-                                "name": "bmi_c",
-                                "model_type_name": "CFE",
-                                "main_output_variable": "Q_OUT",
-                                "init_config": "./config/cat_config/CFE/{{id}}.ini",
-                                "allow_exceed_end_time": true,
-                                "fixed_time_step": false,
-                                "uses_forcing_file": false,
-                                "registration_function": "register_bmi_cfe",
-                                "variables_names_map": {
-                                    "water_potential_evaporation_flux": "EVAPOTRANS",
-                                    "atmosphere_water__liquid_equivalent_precipitation_rate": "raim",
-                                    "ice_fraction_schaake": "sloth_ice_fraction_schaake",
-                                    "ice_fraction_xinanjiang": "sloth_ice_fraction_xinanjiang",
-                                    "soil_moisture_profile": "sloth_soil_moisture_profile"
-                                },
-                                "library_file": "/dmod/shared_libs/libcfebmi.so.1.0.0"
-                            }
-                        }
-                    ],
-                    "uses_forcing_file": false
-                }
-            }
-        ],
-        "forcing": {
-            "path": "./forcings/forcings.nc",
-            "provider": "NetCDF",
-            "enable_cache": false
-        }
-    },
-    "time": {
-        "start_time": "2015-12-01 00:00:00",
-        "end_time": "2015-12-30 23:00:00",
-        "output_interval": 3600
-    },
-    "routing": {
-        "t_route_config_file_with_path": "./config/troute.yaml"
-    },
-    "output_root": "./outputs/ngen"
-}
