+++ tests/test_modular_realization.py
@@ -1,21 +1,4 @@
-"""Tests for ``data_processing.modular_realization``.
-
-The headline test is ``TestGoldenEquivalence`` -- it builds a sloth->nom->cfe
-modular realization and asserts it is identical to the committed
-``tests/golden/realization/cfe-nom.json`` (the hand-authored realization the
-legacy builder produces). That golden is the contract: the modular pipeline
-should reconstruct it exactly.
-
-The remaining classes characterize the individual pieces (validation rules, the
-per-model append/insert helpers, and the integration wiring) so a refactor can't
-drift silently.
-
-Hermeticity: tests that touch output paths monkeypatch ``FilePaths.get_working_dir``
-to a ``tmp_path`` (same pattern as ``test_config_generation.py``) and ``Prompt.ask``
-is always stubbed so nothing blocks on stdin. ``create_modular_realization`` now
-stamps time with ``datetime.strftime`` and therefore expects ``datetime`` inputs,
-so START/END are datetimes (matching ``test_realization_templates.py``).
-"""
+"""Tests for ``data_processing.create_realization``."""
 
 import copy
 import difflib
@@ -26,12 +9,9 @@ from pathlib import Path
 import pytest
 
 from data_processing.file_paths import FilePaths
-import data_processing.modular_realization as mr
-from data_processing.modular_realization import (
-    ACCEPTED_MODELS,
+from data_processing.create_realization import (
     ALL_SLOTH_MODEL_PARAMS,
-    ALL_VARIABLES_NAMES_MAPS,
-    MAIN_OUTPUT_VARIABLES,
+    MODEL_REGISTRY,
     _insert_sloth_module,
     create_modular_realization,
     validate_models,
@@ -47,39 +27,16 @@ CFE_NOM_GOLDEN = GOLDEN_DIR / "cfe-nom.json"
 # ---------------------------------------------------------------------------
 # fixtures / helpers
 # ---------------------------------------------------------------------------
-@pytest.fixture(name="answer_prompt")
-def answer_prompt_fixture(monkeypatch):
-    """Stub ``Prompt.ask``; record whether it was shown and what it returned."""
-
-    class Recorder:  # pylint: disable=too-few-public-methods
-        """Simple prompt-answer recorder used to stub Prompt.ask in tests."""
-
-        called = False
-        response = "n"
-
-        def __call__(self, response):
-            self.response = response
-            return self
-
-    rec = Recorder()
-
-    def fake_ask(*args, **kwargs):  # pylint: disable=unused-argument
-        rec.called = True
-        return rec.response
-
-    monkeypatch.setattr(mr.Prompt, "ask", staticmethod(fake_ask))
-    return rec
 
 
 @pytest.fixture(name="make_realization")
-def make_realization_fixture(tmp_path, monkeypatch, answer_prompt):
+def make_realization_fixture(tmp_path, monkeypatch):
     """Return a runner for ``create_modular_realization`` rooted at ``tmp_path``.
 
     Pre-creates ``config/`` because the function writes ``realization.json`` into
     it but does not create it (the real workflow makes it during subsetting).
     """
     monkeypatch.setattr(FilePaths, "get_working_dir", classmethod(lambda cls: Path(tmp_path)))
-    answer_prompt("y")
 
     def _run(  # pylint: disable=too-many-arguments
         models, *, folder="cat-test", start=START, end=END, routing=False, make_config=True
@@ -87,7 +44,7 @@ def make_realization_fixture(tmp_path, monkeypatch, answer_prompt):
         paths = FilePaths(folder)
         if make_config:
             paths.config_dir.mkdir(parents=True, exist_ok=True)
-        create_modular_realization(folder, start, end, models, routing)
+        create_modular_realization(folder, start, end, models, routing=routing)
         return json.loads((paths.config_dir / "realization.json").read_text())
 
     return _run
@@ -152,31 +109,27 @@ def test_modular_realization_matches_golden(models, golden_name, label, make_rea
 class TestValidateModelsInputs:
     """Validate input handling for the model-selection parser."""
 
-    def test_empty_list_raises(self, answer_prompt):
+    def test_empty_list_raises(self):
         """Ensure an empty model list raises a clear validation error."""
         with pytest.raises(ValueError, match="No models specified"):
-            validate_models([], routing=False)
-        assert not answer_prompt.called
+            _ = validate_models([], routing=False)
 
-    def test_unknown_model_raises_and_names_offender(self, answer_prompt):
+    def test_unknown_model_raises_and_names_offender(
+        self,
+    ):
         """Unknown model names are rejected before any prompt, and the error
         message names the offending model."""
-        with pytest.raises(ValueError) as exc:
-            validate_models(["cfe", "not_a_model"], routing=False)
-        message = str(exc.value)
-        assert "Invalid models specified" in message
-        assert "not_a_model" in message
-        assert not answer_prompt.called
+        with pytest.raises(ValueError, match="Invalid models specified"):
+            _ = validate_models(["cfe", "not_a_model"], routing=False)
 
-    def test_every_accepted_model_is_a_valid_name(self, answer_prompt):
+    def test_every_accepted_model_is_a_valid_name(self):
         """[model] alone must never trip the 'invalid name' guard (it may still
         warn about dependencies -- we answer 'y')."""
-        answer_prompt("y")
-        for model in ACCEPTED_MODELS:
-            try:
-                validate_models([model], routing=False)
-            except ValueError as e:
-                assert "Invalid models specified" not in str(e), model
+        for model in list(MODEL_REGISTRY.keys()):
+            warnings = validate_models([model], routing=False)
+
+        for warning in warnings:
+            assert "Invalid models specified" not in warning
 
 
 # ---------------------------------------------------------------------------
@@ -188,54 +141,27 @@ class TestValidateModelsDependencies:
     WARNING_CASES = [
         (["cfe"], "CFE requires SLoTH"),
         (["casam"], "CASAM requires SLoTH"),
-        (["sft"], "SFT requires SLoTH"),
-        (["smp"], "SMP requires SLoTH"),
-        (["topmodel"], "TOPMODEL requires SLoTH, NOM, or PET"),
         (["sac-sma"], "SAC-SMA requires SLoTH, NOM, or PET"),
     ]
-
-    @pytest.mark.parametrize("models, substring", WARNING_CASES)
-    def test_unmet_dependency_warns_and_aborts_on_no(self, models, substring, answer_prompt):
-        """Ensure a dependency warning aborts when the user declines to proceed."""
-        answer_prompt("n")
-        with pytest.raises(ValueError) as exc:
-            validate_models(models, routing=False)
-        assert answer_prompt.called
-        assert substring in str(exc.value)
-
-    @pytest.mark.parametrize("models", [case[0] for case in WARNING_CASES])
-    def test_unmet_dependency_proceeds_on_yes(self, models, answer_prompt):
-        """Ensure the user can proceed past a dependency warning by answering yes."""
-        answer_prompt("y")
-        validate_models(models, routing=False)  # must not raise
-        assert answer_prompt.called
 
     @pytest.mark.parametrize(
         "models",
         [
             ["sloth", "cfe"],
             ["sloth", "casam"],
-            ["sloth", "sft"],
-            ["sloth", "smp"],
-            ["nom", "topmodel"],
-            ["pet", "topmodel"],
             ["nom", "sac-sma"],
-            ["pet", "sac-sma"],
         ],
     )
-    def test_met_dependency_does_not_prompt(self, models, answer_prompt):
+    def test_met_dependency_does_not_prompt(self, models):
         """Ensure satisfied dependencies do not trigger a confirmation prompt."""
-        validate_models(models, routing=False)
-        assert not answer_prompt.called
+        warnings = validate_models(models, routing=False)
+        assert not warnings
 
-    def test_multiple_unmet_dependencies_accumulate(self, answer_prompt):
+    def test_multiple_unmet_dependencies_accumulate(self):
         """Each failing model contributes its own warning line to the message."""
-        answer_prompt("n")
-        with pytest.raises(ValueError) as exc:
-            validate_models(["cfe", "casam"], routing=False)
-        message = str(exc.value)
-        assert "CFE requires SLoTH" in message
-        assert "CASAM requires SLoTH" in message
+        warnings = validate_models(["cfe", "casam"], routing=False)
+        assert "CFE requires SLoTH" in warnings
+        assert "CASAM requires SLoTH" in warnings
 
 
 # ---------------------------------------------------------------------------
@@ -244,22 +170,20 @@ class TestValidateModelsDependencies:
 class TestValidateModelsRouting:
     """Validate routing-related model selection rules."""
 
-    def test_routing_without_rainfall_runoff_warns(self, answer_prompt):
+    def test_routing_without_rainfall_runoff_warns(self):
         """Ensure routing requires a rainfall-runoff model and prompts otherwise."""
-        answer_prompt("n")
-        with pytest.raises(ValueError, match="Routing is on but no rainfall-runoff"):
-            validate_models(["nom"], routing=True)
-        assert answer_prompt.called
+        warnings = validate_models(["nom"], routing=True)
+        assert "Routing is on but no rainfall-runoff model is used" in warnings
 
-    def test_routing_with_rainfall_runoff_is_quiet(self, answer_prompt):
+    def test_routing_with_rainfall_runoff_is_quiet(self):
         """Ensure routing succeeds without prompting when a runoff model is present."""
-        validate_models(["sloth", "cfe"], routing=True)
-        assert not answer_prompt.called
+        warnings = validate_models(["sloth", "cfe"], routing=True)
+        assert not warnings
 
-    def test_routing_off_never_adds_routing_warning(self, answer_prompt):
+    def test_routing_off_never_adds_routing_warning(self):
         """Ensure routing is ignored when routing is disabled."""
-        validate_models(["nom"], routing=False)
-        assert not answer_prompt.called
+        warnings = validate_models(["nom"], routing=False)
+        assert not warnings
 
 
 # ---------------------------------------------------------------------------
@@ -280,7 +204,7 @@ class TestInsertSlothModule:
     def test_builds_model_params_from_sloth_prefixed_vars(self):
         """Ensure SLOTH parameters are derived from sloth-prefixed variable names."""
         modules = []
-        target = {"cfe": copy.deepcopy(ALL_VARIABLES_NAMES_MAPS["cfe"])}
+        target = {"cfe": copy.deepcopy(MODEL_REGISTRY["cfe"].variables_names_map)}
         _insert_sloth_module(["sloth"], target, modules)
         params = modules[0]["params"]["model_params"]
         assert set(params.keys()) == {
@@ -325,7 +249,9 @@ class TestCreateModularRealization:
         """Ensure the main output variable follows the last model in the chain."""
         r = make_realization(["sloth", "cfe"])
         params = r["global"]["formulations"][0]["params"]
-        assert params["main_output_variable"] == MAIN_OUTPUT_VARIABLES["cfe"] == "Q_OUT"
+        assert (
+            params["main_output_variable"] == MODEL_REGISTRY["cfe"].main_output_variable == "Q_OUT"
+        )
 
     def test_sloth_inserted_before_cfe_by_position(self, make_realization):
         """Ensure the SLOTH module appears before the CFE module in the chain."""
@@ -347,7 +273,11 @@ class TestCreateModularRealization:
         r = make_realization(["sloth", "nom", "casam"])
         assert "CASAM" in _model_type_names(r)
         params = r["global"]["formulations"][0]["params"]
-        assert params["main_output_variable"] == MAIN_OUTPUT_VARIABLES["casam"] == "total_discharge"
+        assert (
+            params["main_output_variable"]
+            == MODEL_REGISTRY["casam"].main_output_variable
+            == "total_discharge"
+        )
 
     def test_nom_override_rewrites_casam_pet_source(self, make_realization):
         """nom seen before casam -> casam's PET source switches to the Noah-OWP
@@ -362,9 +292,8 @@ class TestCreateModularRealization:
         """cfe before nom -> cfe keeps its default (sloth/forcing) sources."""
         r = make_realization(["sloth", "cfe", "nom"])
         vmap = _cfe_module(r)["params"]["variables_names_map"]
-        assert (
-            vmap["water_potential_evaporation_flux"]
-            == (ALL_VARIABLES_NAMES_MAPS["cfe"]["water_potential_evaporation_flux"])
+        assert vmap["water_potential_evaporation_flux"] == (
+            MODEL_REGISTRY["cfe"].variables_names_map["water_potential_evaporation_flux"]
         )
 
     def test_routing_on_adds_troute_block(self, make_realization):
