+++ tests/test_summa_config_generation.py
@@ -2,7 +2,6 @@
 
 import difflib
 import json
-import shutil
 from datetime import datetime
 from pathlib import Path
 
@@ -11,14 +10,15 @@ import pandas as pd
 import pytest
 import xarray as xr
 
-from data_processing.create_realization import (
-    get_hru_order,
-    make_summa_attributes,
-    make_summa_trialParams,
-    make_summa_coldState,
-    make_summa_config,
+from data_processing.create_configs import (  # pylint: disable=no-name-in-module
+    _get_hru_order as get_hru_order,
+    _make_summa_attributes as make_summa_attributes,
+    _make_summa_trialParams as make_summa_trialParams,
+    _make_summa_coldState as make_summa_coldState,
+    _make_summa_config_suite as make_summa_config_suite,
 )
 from data_processing.file_paths import FilePaths
+from golden_utils import SUMMA_GAGE_FORCING_IDS as GAGE_FORCING_IDS
 
 GOLDEN_GPKG_DIR = Path(__file__).parent / "golden" / "geopackage"
 GOLDEN_CONFIG_DIR = Path(__file__).parent / "golden" / "config"
@@ -198,96 +198,6 @@ EXPECTED_SUMMA_TRIAL_PARAMS = {
     },
 }
 
-GAGE_FORCING_IDS = [
-    "cat-2861379",
-    "cat-2861380",
-    "cat-2861387",
-    "cat-2861414",
-    "cat-2861421",
-    "cat-2861429",
-    "cat-2861431",
-    "cat-2861436",
-    "cat-2861438",
-    "cat-2861442",
-    "cat-2861446",
-    "cat-2861447",
-    "cat-2861449",
-    "cat-2861452",
-    "cat-2861453",
-    "cat-2861471",
-    "cat-2861472",
-    "cat-2861475",
-    "cat-2861488",
-    "cat-2861382",
-    "cat-2861383",
-    "cat-2861384",
-    "cat-2861385",
-    "cat-2861388",
-    "cat-2861389",
-    "cat-2861391",
-    "cat-2861419",
-    "cat-2861420",
-    "cat-2861423",
-    "cat-2861427",
-    "cat-2861430",
-    "cat-2861433",
-    "cat-2861439",
-    "cat-2861457",
-    "cat-2861458",
-    "cat-2861468",
-    "cat-2861477",
-    "cat-2861478",
-    "cat-2861485",
-    "cat-2861474",
-    "cat-2861487",
-    "cat-2861476",
-    "cat-2861486",
-    "cat-2861484",
-    "cat-2861480",
-    "cat-2861482",
-    "cat-2861481",
-    "cat-2861483",
-    "cat-2861479",
-    "cat-2861473",
-    "cat-2861470",
-    "cat-2861469",
-    "cat-2861467",
-    "cat-2861466",
-    "cat-2861381",
-    "cat-2861462",
-    "cat-2861464",
-    "cat-2861463",
-    "cat-2861465",
-    "cat-2861461",
-    "cat-2861459",
-    "cat-2861460",
-    "cat-2861455",
-    "cat-2861456",
-    "cat-2861454",
-    "cat-2861451",
-    "cat-2861450",
-    "cat-2861448",
-    "cat-2861445",
-    "cat-2861444",
-    "cat-2861441",
-    "cat-2861443",
-    "cat-2861440",
-    "cat-2861437",
-    "cat-2861435",
-    "cat-2861434",
-    "cat-2861432",
-    "cat-2861386",
-    "cat-2861428",
-    "cat-2861426",
-    "cat-2861425",
-    "cat-2861424",
-    "cat-2861390",
-    "cat-2861422",
-    "cat-2861415",
-    "cat-2861417",
-    "cat-2861418",
-    "cat-2861416",
-]
 GAGE_HRU_IDS = [int(gid.split("-")[1]) for gid in GAGE_FORCING_IDS]
 
 EXPECTED_GAGE_SUMMA_ATTRIBUTE_FIRST_LAST = {
@@ -380,47 +290,6 @@ def _assert_dataset_matches_expected(ds: xr.Dataset, expected: dict, rtol=1e-9, 
                 )
             else:
                 assert np.array_equal(got_arr, want_arr), f"{name}: {got_arr} != {want_arr}"
-
-
-def _generate_summa_config(cat_id: str, forcing_path: Path, tmp_root: Path, monkeypatch) -> dict:
-    monkeypatch.setattr(FilePaths, "get_working_dir", classmethod(lambda cls: Path(tmp_root)))
-    output_dir = Path(tmp_root) / "config"
-
-    summa_model_config = output_dir / "model_config" / "SUMMA"
-    summa_model_config.mkdir(parents=True, exist_ok=True)
-
-    for path in FilePaths.summa_file_dir.glob("*"):
-        if not path.is_file() or path.suffix == ".nc":
-            continue
-        if path.name == "fileManager.txt":
-            template = path.read_text()
-            (summa_model_config / path.name).write_text(
-                template.format(
-                    start_time=FIXED_START.strftime("%Y-%m-%d %H:%M:%S"),
-                    end_time=FIXED_END.strftime("%Y-%m-%d %H:%M:%S"),
-                )
-            )
-        else:
-            shutil.copy(path, summa_model_config / path.name)
-
-    hru_ids = get_hru_order(forcing_path)
-    make_summa_attributes(hru_ids, GEOPACKAGE_FIXTURES[cat_id]).to_netcdf(
-        summa_model_config / "attributes.nc"
-    )
-    make_summa_trialParams(
-        hru_ids, int((FIXED_END - FIXED_START).total_seconds() / 3600)
-    ).to_netcdf(summa_model_config / "trialParams.nc")
-    ds, encoding = make_summa_coldState(hru_ids)
-    ds.to_netcdf(summa_model_config / "coldState.nc", encoding=encoding)
-    make_summa_config(hru_ids, output_dir)
-
-    produced = {}
-    for f in sorted(output_dir.rglob("*")):
-        if f.is_file() and f.suffix != ".nc":
-            produced[str(f.relative_to(output_dir))] = _normalize(
-                f.read_text(encoding="utf-8", errors="replace"), output_dir
-            )
-    return produced
 
 
 @pytest.fixture(name="cat_1555522_forcing_output")
@@ -524,62 +393,6 @@ def test_make_summa_trialParams_netcdf_matches_expected(tmp_path):  # pylint: di
         _assert_dataset_matches_expected(actual_ds, EXPECTED_SUMMA_TRIAL_PARAMS)
 
 
-def test_summa_config_generation_matches_golden(cat_1555522_forcing_output, tmp_path, monkeypatch):
-    """Checks all non-netCDF config files."""
-    produced = _generate_summa_config(
-        "cat-1555522", cat_1555522_forcing_output["forcings_nc"], tmp_path, monkeypatch
-    )
-    assert GOLDEN_SUMMA_FILE.exists(), (
-        f"missing golden {GOLDEN_SUMMA_FILE}. Generate it with: "
-        "UPDATE_GOLDEN=1 uv run pytest tests/test_summa_config_generation.py"
-    )
-
-    golden = json.loads(GOLDEN_SUMMA_FILE.read_text())
-
-    missing = sorted(set(golden) - set(produced))
-    extra = sorted(set(produced) - set(golden))
-    assert not missing and not extra, (
-        f"SUMMA config file set changed.\n  missing: {missing}\n  extra: {extra}"
-    )
-
-    changed = [p for p in sorted(golden) if golden[p] != produced[p]]
-    if changed:
-        first = changed[0]
-        diff = "\n".join(
-            difflib.unified_diff(
-                golden[first].splitlines(),
-                produced[first].splitlines(),
-                fromfile=f"golden/{first}",
-                tofile=f"produced/{first}",
-                lineterm="",
-            )
-        )
-        pytest.fail(
-            f"{len(changed)} SUMMA config file(s) changed: {changed}\nfirst diff ({first}):\n{diff}"
-        )
-
-
-def test_summa_config_generation_produces_expected_artifacts(
-    cat_1555522_forcing_output, tmp_path, monkeypatch
-):
-    """Checks that all non-netCDF files exist."""
-    produced = _generate_summa_config(
-        "cat-1555522", cat_1555522_forcing_output["forcings_nc"], tmp_path, monkeypatch
-    )
-    assert "cat_config/SUMMA/cat-1555522.input" in produced
-    assert "model_config/SUMMA/fileManager.txt" in produced
-    assert "model_config/SUMMA/README.md" in produced
-    assert "model_config/SUMMA/forcingFileList.txt" in produced
-    assert "model_config/SUMMA/basinParamInfo.txt" in produced
-    assert "model_config/SUMMA/localParamInfo.txt" in produced
-    assert "model_config/SUMMA/modelDecisions.txt" in produced
-    assert "model_config/SUMMA/outputControl.txt" in produced
-    assert "model_config/SUMMA/TBL_GENPARM.TBL" in produced
-    assert "model_config/SUMMA/TBL_MPTABLE.TBL" in produced
-    assert "model_config/SUMMA/TBL_SOILPARM.TBL" in produced
-    assert "model_config/SUMMA/TBL_VEGPARM.TBL" in produced
-
-
 def test_gage_get_hru_order_returns_expected_ids(gage_10109001_forcing_output):
     """Checks get_hru_order for a multi-catchment simulation."""
     ids = get_hru_order(gage_10109001_forcing_output["forcings_nc"])
@@ -640,23 +453,53 @@ def test_gage_make_summa_trialParams_dims_and_structure(tmp_path):  # pylint: di
         assert int(actual["hruId"].values[-1]) == GAGE_HRU_IDS[-1]
         np.testing.assert_allclose(actual["maxstep"].values, timesteps * 3600)
 
+# ---------------------------------------------------------------------------
+# Integration tests for _make_summa_config_suite.
+#
+# The tests above exercise the SUMMA leaf functions (make_summa_attributes,
+# make_summa_coldState, make_summa_trialParams, make_summa_config) and compare
+# text config against the golden files, but they do so through the local
+# _generate_summa_config helper, which *reimplements* the orchestration. The
+# real modular entry point, _make_summa_config_suite (create_configs.py),
+# is never invoked there -- so an orchestration bug in it (wrong glob, wrong
+# hydrofabric source, a dropped setup_run_folders, or fileManager templating
+# drift) would pass unnoticed. These tests call the real function so the
+# three copies of the assembly logic cannot silently diverge.
+# ---------------------------------------------------------------------------
 
-def test_gage_summa_config_generation_matches_golden(
-    gage_10109001_forcing_output, tmp_path, monkeypatch
-):
-    """Checks that the multi-catchment non-netCDF files match the reference."""
-    produced = _generate_summa_config(
-        "gage-10109001", gage_10109001_forcing_output["forcings_nc"], tmp_path, monkeypatch
-    )
-    assert GOLDEN_GAGE_SUMMA_FILE.exists(), (
-        f"missing golden {GOLDEN_GAGE_SUMMA_FILE}. Save the gage golden there."
-    )
-    golden = json.loads(GOLDEN_GAGE_SUMMA_FILE.read_text())
+
+def _run_summa_config_suite(cat_id: str, gpkg: Path, monkeypatch) -> tuple[dict, FilePaths]:
+    """Invokes the real make_summa_config_suite and returns (produced, paths).
+
+    Assumes the caller's forcing fixture has already pointed get_working_dir at
+    the tmp working dir and written forcings.nc. conus_hydrofabric is redirected
+    to the subset geopackage fixture -- make_summa_config_suite reads attributes
+    straight off FilePaths.conus_hydrofabric, which is not shippable in tests.
+    """
+    monkeypatch.setattr(FilePaths, "conus_hydrofabric", gpkg)
+
+    make_summa_config_suite(cat_id, FIXED_START, FIXED_END)
+
+    paths = FilePaths(cat_id)
+    output_dir = paths.config_dir
+    produced = {}
+    for f in sorted(output_dir.rglob("*")):
+        if f.is_file() and f.suffix != ".nc":
+            produced[str(f.relative_to(output_dir))] = _normalize(
+                f.read_text(encoding="utf-8", errors="replace"), output_dir
+            )
+    return produced, paths
+
+
+def _assert_matches_golden(produced: dict, golden_file: Path, label: str) -> None:
+    """Shared golden-set comparison for the config-suite tests."""
+    assert golden_file.exists(), f"missing golden {golden_file}"
+    golden = json.loads(golden_file.read_text())
 
     missing = sorted(set(golden) - set(produced))
     extra = sorted(set(produced) - set(golden))
     assert not missing and not extra, (
-        f"SUMMA config file set changed for gage-10109001.\n  missing: {missing}\n  extra: {extra}"
+        f"SUMMA config file set changed for {label}.\n  missing: {missing}\n  extra: {extra}"
     )
 
     changed = [p for p in sorted(golden) if golden[p] != produced[p]]
@@ -672,33 +515,69 @@ def test_gage_summa_config_generation_matches_golden(
             )
         )
         pytest.fail(
-            f"{len(changed)} SUMMA config file(s) changed for gage-10109001: {changed}\n"
+            f"{len(changed)} SUMMA config file(s) changed for {label}: {changed}\n"
             f"first diff ({first}):\n{diff}"
         )
 
 
-def test_gage_summa_config_generation_produces_expected_artifacts(
-    gage_10109001_forcing_output, tmp_path, monkeypatch
-):
-    """Checks that all non-netCDF files for a multi-catchment simulation were generated."""
-    produced = _generate_summa_config(
-        "gage-10109001", gage_10109001_forcing_output["forcings_nc"], tmp_path, monkeypatch
+def test_summa_config_suite_matches_golden(cat_1555522_forcing_output, monkeypatch):  # pylint: disable=unused-argument
+    """The suite produces the same non-netCDF config files as the local helper."""
+    produced, _ = _run_summa_config_suite(
+        "cat-1555522", GEOPACKAGE_FIXTURES["cat-1555522"], monkeypatch
     )
+    _assert_matches_golden(produced, GOLDEN_SUMMA_FILE, "cat-1555522")
+
+
+def test_summa_config_suite_writes_expected_netcdf(cat_1555522_forcing_output, monkeypatch):  # pylint: disable=unused-argument
+    """The suite wires the netCDF writers to summa_model_config with correct inputs."""
+    _, paths = _run_summa_config_suite(
+        "cat-1555522", GEOPACKAGE_FIXTURES["cat-1555522"], monkeypatch
+    )
+    model_config = paths.summa_model_config
+
+    with xr.open_dataset(model_config / "attributes.nc") as ds:
+        _assert_dataset_matches_expected(ds, EXPECTED_SUMMA_ATTRIBUTES)
+    with xr.open_dataset(model_config / "coldState.nc") as ds:
+        _assert_dataset_matches_expected(ds, EXPECTED_SUMMA_COLD_STATE)
+    with xr.open_dataset(model_config / "trialParams.nc") as ds:
+        _assert_dataset_matches_expected(ds, EXPECTED_SUMMA_TRIAL_PARAMS)
+
+
+def test_summa_config_suite_sets_up_run_folders(cat_1555522_forcing_output, monkeypatch):  # pylint: disable=unused-argument
+    """The suite must call setup_run_folders(['outputs/summa']); the helper never did."""
+    _, paths = _run_summa_config_suite(
+        "cat-1555522", GEOPACKAGE_FIXTURES["cat-1555522"], monkeypatch
+    )
+    for folder in ("outputs/summa", "outputs/ngen", "outputs/troute", "metadata"):
+        assert (paths.subset_dir / folder).is_dir(), f"missing run folder: {folder}"
+
+
+def test_summa_config_suite_templates_file_manager(cat_1555522_forcing_output, monkeypatch):  # pylint: disable=unused-argument
+    """fileManager.txt is templated with the run dates and no placeholders remain.
+
+    make_summa_config_suite formats the template with raw datetime objects, so
+    this also pins the str(datetime) rendering the golden was generated against.
+    """
+    produced, _ = _run_summa_config_suite(
+        "cat-1555522", GEOPACKAGE_FIXTURES["cat-1555522"], monkeypatch
+    )
+    file_manager = produced["model_config/SUMMA/fileManager.txt"]
+    assert "{start_time}" not in file_manager
+    assert "{end_time}" not in file_manager
+    assert f"simStartTime         '{FIXED_START}'" in file_manager
+    assert f"simEndTime           '{FIXED_END}'" in file_manager
+
+
+def test_gage_summa_config_suite_matches_golden(gage_10109001_forcing_output, monkeypatch):  # pylint: disable=unused-argument
+    """The multi-catchment gage config assembled by the real suite matches golden."""
+    produced, paths = _run_summa_config_suite(
+        "gage-10109001", GEOPACKAGE_FIXTURES["gage-10109001"], monkeypatch
+    )
+    _assert_matches_golden(produced, GOLDEN_GAGE_SUMMA_FILE, "gage-10109001")
+
     cat_inputs = [k for k in produced if k.startswith("cat_config/SUMMA/") and k.endswith(".input")]
     assert len(cat_inputs) == len(GAGE_HRU_IDS)
-    assert "cat_config/SUMMA/cat-2861379.input" in produced  # first in HRU order
-    assert "cat_config/SUMMA/cat-2861416.input" in produced  # last in HRU order
-    for name in (
-        "fileManager.txt",
-        "README.md",
-        "forcingFileList.txt",
-        "basinParamInfo.txt",
-        "localParamInfo.txt",
-        "modelDecisions.txt",
-        "outputControl.txt",
-        "TBL_GENPARM.TBL",
-        "TBL_MPTABLE.TBL",
-        "TBL_SOILPARM.TBL",
-        "TBL_VEGPARM.TBL",
-    ):
-        assert f"model_config/SUMMA/{name}" in produced
+    with xr.open_dataset(paths.summa_model_config / "attributes.nc") as ds:
+        assert ds.sizes["hru"] == len(GAGE_HRU_IDS)
+        assert int(ds["hruId"].values[0]) == GAGE_HRU_IDS[0]
+        assert int(ds["hruId"].values[-1]) == GAGE_HRU_IDS[-1]
